* tuning

+++ src/main/java/yk/senjin/shaders/gshader/examples/blend/BlendF.groovy
@@ -19,7 +19,7 @@ class BlendF extends FragmentShaderParent<BlendFi> {
 
     //TODO fix without initialization ProgramGenerator fails
     public Sampler2D txt = new Sampler2D(); //размываемая текстура
-    public float kSize; //размер ядра
+    public int kSize; //размер ядра
     public float[] koeff = new float[MAX_KOEFF_SIZE]; //коэффициенты
     public Vec2f direction = new Vec2f(0.003f, 0f); //направление размытия с учетом радиуса размытия и aspect ratio, например (0.003, 0.0) - горизонтальное и (0.0, 0.002) - вертикальное
